# server/pkg/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package db

import (
	"github.com/google/uuid"
	"github.com/jackc/pgx/v5/pgtype"
)

type Link struct {
	ID          uuid.UUID        `json:"id"`
	Shortcode   string           `json:"shortcode"`
	OriginalUrl string           `json:"original_url"`
	UserID      string           `json:"user_id"`
	Clicks      *int32           `json:"clicks"`
	ExpiresAt   pgtype.Timestamp `json:"expires_at"`
	CreatedAt   pgtype.Timestamp `json:"created_at"`
	UpdatedAt   pgtype.Timestamp `json:"updated_at"`
	DeletedAt   pgtype.Timestamp `json:"deleted_at"`
}